// frontend/src/types/generated/EmailTriageInboxSettingsResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { EmailTriageGmailCategory } from "./EmailTriageGmailCategory";

export type EmailTriageInboxSettingsResponse = { revision: number | null, visible_gmail_categories: Array<EmailTriageGmailCategory>, };